added definition command, tidied up some code
Definition command will be used for defining things that are not a role. For example if I wanted to know the definition of objectivism I could use /define objectivism. I also changed all instances of ctx.message.content[8:] to a variable named label in order to make the code look cleaner.